+++ Gravitas.Model/DomainModel/PreRegistration/DAO/PreRegisterProduct.cs
@@ -1,12 +1,13 @@
 using System.ComponentModel.DataAnnotations.Schema;
 using Gravitas.Model.DomainModel.Base;
+using Gravitas.Model.DomainModel.PredefinedRoute.DAO;
 
 namespace Gravitas.Model.DomainModel.PreRegistration.DAO
 {
     [Table("PreRegisterProduct")]
     public class PreRegisterProduct : BaseEntity<int>
     {
-        public long RouteTemplateId { get; set; }
+        public int RouteTemplateId { get; set; }
         public string Title { get; set; }
         public int RouteTimeInMinutes { get; set; }
