+++ Gravitas.Model/DomainModel/PreRegistration/DAO/PreRegisterQueue.cs
@@ -1,6 +1,7 @@
 using System;
 using System.ComponentModel.DataAnnotations.Schema;
 using Gravitas.Model.DomainModel.Base;
+using Gravitas.Model.DomainModel.PredefinedRoute.DAO;
 
 namespace Gravitas.Model.DomainModel.PreRegistration.DAO
 {
@@ -8,8 +9,8 @@ namespace Gravitas.Model.DomainModel.PreRegistration.DAO
     public class PreRegisterQueue : BaseEntity<int>
     {
         public string PhoneNo { get; set; }
-        public long PreRegisterCompanyId { get; set; }
-        public long RouteTemplateId { get; set; }
+        public int PreRegisterCompanyId { get; set; }
+        public int RouteTemplateId { get; set; }
         public DateTime RegisterDateTime { get; set; }
         public string TruckNumber { get; set; }
         public string Notice { get; set; }
